requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aistudio-sdk==0.3.6
alibabacloud-bailian20230601==1.6.1
alibabacloud-credentials==1.0.1
alibabacloud-credentials-api==1.0.0
alibabacloud-tea==0.4.3
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.14
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
APScheduler==3.10.4
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.0
backports.zoneinfo==0.2.1
bce-python-sdk==0.9.45
beautifulsoup4==4.13.4
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.8
colorlog==6.9.0
contourpy==1.1.1
cryptography==44.0.2
cycler==0.12.1
decorator==5.1.1
effdet==0.4.1
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hf-xet==1.1.5
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.33.4
hyperframe==5.2.0
idna==2.10
imagesize==1.4.1
importlib-metadata==6.0.0
importlib_resources==6.4.5
iopath==0.1.10
Jinja2==3.1.6
joblib==1.2.0
jplephem==2.23
kiwisolver==1.4.7
layoutparser==0.3.4
librosa==0.9.2
llvmlite==0.31.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
narwhals==1.42.1
networkx==3.1
numba==0.48.0
numpy==1.24.4
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==23.0
paddlepaddle==3.0.0
pandas==2.0.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
pooch==1.6.0
portalocker==3.0.0
prettytable==3.11.0
propcache==0.2.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.28.2
resampy==0.3.1
rfc3986==1.5.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.8
safetensors==0.5.3
scikit-learn==1.2.1
scipy==1.10.0
sgp4==2.25
shapely==2.0.7
six==1.16.0
skyfield==1.53
sniffio==1.3.1
soundfile==0.11.0
soupsieve==2.7
sympy==1.13.3
threadpoolctl==3.1.0
timm==1.0.17
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.14
wcwidth==0.2.13
yarl==1.15.2
zipp==3.12.0
